Refactored petfinder's types
