+++ src/lib/petfinderTypes.d.ts
